completed refactor of unit tests
