+++ test/stim/passes/stim_reference_programs/qubit/qubit.txt
@@ -1,0 +1,5 @@
+
+H 0 1
+X 0
+CX 1 0
+MZ(0.00000000) 0 1
