+++ test/stim/passes/stim_reference_programs/qubit/qubit_broadcast.txt
@@ -1,0 +1,3 @@
+
+H 0 1 2 3
+MZ(0.00000000) 0 1 2 3
