+++ test/stim/passes/stim_reference_programs/qubit/qubit_reset.txt
@@ -1,0 +1,3 @@
+
+RZ 0
+MZ(0.00000000) 0
