+++ test/stim/passes/stim_reference_programs/wire/wire.txt
@@ -1,0 +1,3 @@
+
+S 0
+H 0 1 2 3
