+++ test/stim/passes/stim_reference_programs/wire/wire_apply.txt
@@ -1,0 +1,2 @@
+
+S 0
