+++ test/stim/passes/stim_reference_programs/wire/wire_apply_control.txt
@@ -1,0 +1,2 @@
+
+CX 0 1
