+++ test/stim/passes/stim_reference_programs/wire/wire_broadcast.txt
@@ -1,0 +1,2 @@
+
+H 0 1 2 3
