+++ test/stim/passes/stim_reference_programs/wire/wire_broadcast_control.txt
@@ -1,0 +1,6 @@
+
+CX 0 1 2 3
+MZ(0.00000000) 0
+MZ(0.00000000) 1
+MZ(0.00000000) 2
+MZ(0.00000000) 3
