+++ test/stim/passes/stim_reference_programs/wire/wire_measure.txt
@@ -1,0 +1,2 @@
+
+MZ(0.00000000) 0
