+++ test/stim/passes/stim_reference_programs/wire/wire_multiple_apply.txt
@@ -1,0 +1,5 @@
+
+S 0
+H 0
+I 0
+I 0
