+++ test/stim/passes/stim_reference_programs/wire/wire_reset.txt
@@ -1,0 +1,3 @@
+
+RZ 0
+MZ(0.00000000) 0
